% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropWhen.R
\name{dropWhen}
\alias{dropWhen}
\title{Remove columns with specific values}
\usage{
dropWhen(df, unique = FALSE, values, requireAll = TRUE)
}
\arguments{
\item{df}{Data frame to column filter}

\item{unique}{Boolean for whether to return only unique rows after column filtering}

\item{values}{Character vector containing all the values besides NA for which a column will be removed}

\item{requireAll}{Boolean as to whether you want to require all rows in a column to equal a value in \code{values} before filtering the column out.}
}
\value{
Returns a columnn filtered data frame
}
\description{
Remove columns with specific values
}
\examples{
df <- data.frame(this = c(NA, seq(1, 10, 1)), that = rep(0, 11), thisotherThing = rep(NA))
cleaner <- dropWhen(df, unique = FALSE, values = c("0"), requireAll = FALSE)
}
\author{
Amy Paguirigan
}
