% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropWhen.R
\name{dropWhen}
\alias{dropWhen}
\title{Remove columns with specific trash}
\usage{
dropWhen(df, unique = FALSE, trash = NULL, requireAll = TRUE)
}
\arguments{
\item{df}{Data frame to column filter}

\item{unique}{Boolean for whether to return only unique rows after column filtering}

\item{trash}{Character vector containing all the trash values, besides NA, for which a column can be removed}

\item{requireAll}{Boolean as to whether you want to require all rows in a column to equal a value in \code{trash} before filtering the column out (TRUE).}
}
\value{
Returns a columnn filtered data frame.  Note, does not take into account numeric, integer or character value differences (it will treat them all the same).
}
\description{
Remove columns with specific trash
}
\examples{
df <- data.frame(this = c(NA, seq(1, 5, 1), seq(1, 5, 1)), that = rep(0, 11),
thisotherThing = rep(NA), ohAndThis = rep(c("trash", "0"), 11))
cleanerdf <- dropWhen(df, unique = FALSE, trash = c("0", "trash"), requireAll = FALSE)
}
\author{
Amy Paguirigan
}
